app/requirements.txt
# App dependencies — installed in the Docker container
# Pin major versions to keep the image reproducible

# Web framework
fastapi>=0.111,<1
uvicorn[standard]>=0.29,<1

# LangChain ecosystem
langchain>=0.2,<0.4
langchain-community>=0.2,<0.4

# Embeddings (sentence-transformers, CPU inference in the container)
sentence-transformers>=2.7,<4

# FAISS — CPU build (no GPU required)
faiss-cpu>=1.8,<2

# Ollama client
ollama>=0.4,<1

# Environment variable loading
python-dotenv>=1.0,<2

# Discord bot
discord.py>=2.3,<3
